chore(Validation): rename Enum rule
- Change API and update unit tests
