+++ tests/ValidationTest.php
@@ -4,7 +4,7 @@ namespace NeuronAI\Tests;
 
 use NeuronAI\StructuredOutput\StructuredOutputException;
 use NeuronAI\StructuredOutput\Validation\Rules\ArrayOf;
-use NeuronAI\StructuredOutput\Validation\Rules\Choice;
+use NeuronAI\StructuredOutput\Validation\Rules\Enum;
 use NeuronAI\StructuredOutput\Validation\Rules\Count;
 use NeuronAI\StructuredOutput\Validation\Rules\Email;
 use NeuronAI\StructuredOutput\Validation\Rules\EqualTo;
@@ -24,6 +24,7 @@ use NeuronAI\StructuredOutput\Validation\Rules\IsNotNull;
 use NeuronAI\StructuredOutput\Validation\Rules\Url;
 use NeuronAI\StructuredOutput\Validation\Validator;
 use NeuronAI\Tests\Stubs\DummyEnum;
+use NeuronAI\Tests\Stubs\IntEnum;
 use NeuronAI\Tests\Stubs\Person;
 use NeuronAI\Tests\Stubs\StringEnum;
 use PHPUnit\Framework\TestCase;
@@ -470,14 +471,17 @@ class ValidationTest extends TestCase
         $this->assertCount(1, $violations);
     }
 
-    public function test_choice_validation()
+    public function test_enum_validation()
     {
         $class = new class () {
-            #[Choice(choices: ['one', 'two', 'three'])]
+            #[Enum(values: ['one', 'two', 'three'])]
             public string $number = 'one';
 
-            #[Choice(enum: StringEnum::class)]
+            #[Enum(class: StringEnum::class)]
             public string $enumNumber = 'one';
+
+            #[Enum(class: IntEnum::class)]
+            public IntEnum $intEnum = IntEnum::ONE;
         };
 
         $obj = new $class();
@@ -492,12 +496,14 @@ class ValidationTest extends TestCase
         $obj->enumNumber = 'five';
         $violations = Validator::validate($obj);
         $this->assertCount(2, $violations);
+
+        // $obj->intEnum = 3; -> TypeError anyway...
     }
 
-    public function test_choice_validation_exception_both_option_provided()
+    public function test_enum_validation_exception_both_option_provided()
     {
         $class = new class () {
-            #[Choice(choices: ['one', 'two', 'three'], enum: StringEnum::class)]
+            #[Enum(values: ['one', 'two', 'three'], class: StringEnum::class)]
             public string $number = 'one';
         };
 
@@ -508,10 +514,10 @@ class ValidationTest extends TestCase
         $violations = Validator::validate($obj);
     }
 
-    public function test_choice_validation_exception_no_option_provided()
+    public function test_enum_validation_exception_no_option_provided()
     {
         $class = new class () {
-            #[Choice()]
+            #[Enum()]
             public string $number = 'one';
         };
 
@@ -525,7 +531,7 @@ class ValidationTest extends TestCase
     public function test_choice_validation_exception_invalid_enum()
     {
         $class = new class () {
-            #[Choice(enum: Person::class)]
+            #[Enum(class: Person::class)]
             public string $number = 'one';
         };
 
@@ -536,10 +542,10 @@ class ValidationTest extends TestCase
         $violations = Validator::validate($obj);
     }
 
-    public function test_choice_validation_exception_enum_non_backed()
+    public function test_enum_validation_exception_enum_non_backed()
     {
         $class = new class () {
-            #[Choice(enum: DummyEnum::class)]
+            #[Enum(class: DummyEnum::class)]
             public string $number = 'one';
         };
